+++ src/Metadata/Resource/Factory/ParameterResourceMetadataCollectionFactory.php
@@ -32,6 +32,7 @@ use ApiPlatform\State\Util\StateOptionsTrait;
 use Psr\Container\ContainerInterface;
 use Psr\Log\LoggerInterface;
 use Symfony\Component\Serializer\NameConverter\NameConverterInterface;
+use Symfony\Component\TypeInfo\Type;
 
 /**
  * Prepares Parameters documentation by reading its filter details and declaring an OpenApi parameter.
@@ -148,6 +149,16 @@ final class ParameterResourceMetadataCollectionFactory implements ResourceMetada
             }
 
             $parameter = $this->setDefaults($key, $parameter, $resourceClass, $properties, $operation);
+            // We don't do any type cast yet, a query parameter or an header is always a string or a list of strings
+            if (null === $parameter->getNativeType()) {
+                // this forces the type to be only a list
+                if ('array' === ($parameter->getSchema()['type'] ?? null)) {
+                    $parameter = $parameter->withNativeType(Type::list(Type::string()));
+                } else {
+                    $parameter = $parameter->withNativeType(Type::union(Type::string(), Type::list(Type::string())));
+                }
+            }
+
             $priority = $parameter->getPriority() ?? $internalPriority--;
             $parameters->add($key, $parameter->withPriority($priority));
         }